move QgsGeometryCollection tests from QgsGeometry
